Just for testing pin branches
Long story: we use a composer plugin to force using GitHub releases zip
files. I'm testing a fix to make it work with both composer 1 and 2.
Unfortunately our current setup will first install the base
dependencies, which still requires the broken version of the plugin.
Then on the install of the instance dependencies, it still runs the
broken plugin and crashes the install.